feature(Lab5): внесены изменения, согласно замечаниям

+++ OOP/OOP/Lab5/Product.h
@@ -7,7 +7,7 @@ class Product
 {
 private:
 	string _name;
-	int _price;
+	double _price;
 	CategoryType _category;
 
 public:
